bots/courier_bot/requirements.txt
aiogram>=3.13
python-dotenv
aiohttp
SQLAlchemy>=2.0
asyncpg
